+++ src/components/Navbar.tsx
@@ -1,0 +1,15 @@
+function Navbar() {
+  
+  return(
+    <nav className="site-nav grid">
+      <h1>My Portfolio</h1>
+      <ul>
+        <li><a href="#portfolio">Portfolio</a></li>
+        <li><a href="#skills">Skills</a></li>
+        <li><a href="#contact">Contact</a></li>
+      </ul>
+    </nav>
+  )
+}
+
+export default Navbar