+++ CHANGELOG.md
@@ -10,9 +10,18 @@ project adheres to [Semantic Versioning](https://semver.org).
 - Portfolio website migration to React-Vite version.
 
 ## [1.0.1] - 2024-02-15
-
 - *gitignore*: .vite folder to be ignored.
 
 ## [test1] - 2024-02-17
+
 ## [test2] - 2024-02-17
 - *App.tsx* minimal change in h1
+
+## [1.1.0] - 2024-02-18
+Componentization instead of pure jsx.
+### Updated
+- *App.tsx*: Componentization
+- Folder name: `img` per `icons`.
+- *Main.tsx*: `index.css` import deleted.
+### Added
+- New components: Contact, Footer, Navbar, Portfolio, Skills and Welcome.